% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_pb.R
\name{multi_pb}
\alias{multi_pb}
\title{Multiple progress bars}
\usage{
multi_pb(m, data_list)
}
\arguments{
\item{m}{a CmdStanModel object}

\item{data_list}{a list of data to feed to \link[cmdstanr]{sample}}
}
\description{
Show multiple concurrent progress bars while fitting a model with `cmdstanr`
}
